pyproject.toml
[project]
name = "01-chatbot-gemini"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.129.0",
    "openai-agents>=0.9.1",
    "uvicorn>=0.41.0",
]
